+++ include/flamenco/core/pin.h
@@ -21,7 +21,7 @@ protected:
     
     // Заполняет буферы каналов звуковыми данными.
     // Размер каждого буфера - CHANNEL_BUFFER_SIZE_IN_SAMPLES семплов.
-    virtual void process( s16 * left, s16 * right ) = 0;
+    virtual void process( f32 * left, f32 * right ) = 0;
     friend class Mixer;
     
     // Присоединен ли пин к другому пину или микшеру.
